refactor foobar console file, add `source-code` hint
